Fix guard checks

+++ scripts/drivers/aws-secretsmanager.sh
@@ -32,23 +32,24 @@ _custom_driver_get_secret() {
         fi
     fi
 
-    if [ ! -z ${AWS_ROLE_ARN} || ! -z ${AWS_WEB_IDENTITY_TOKEN_FILE} ]; then
-        aws sts assume-role-with-web-identity \
-	        --role-arn $AWS_ROLE_ARN \
-            --role-session-name x-account \
-            --web-identity-token file://$AWS_WEB_IDENTITY_TOKEN_FILE \
-            --duration 1500 >/tmp/temp_creds.txt
+    if ! aws sts get-caller-identity --output text; then
+        if [ ! -z ${AWS_ROLE_ARN} || ! -z ${AWS_WEB_IDENTITY_TOKEN_FILE} ]; then
+            aws sts assume-role-with-web-identity \
+                --role-arn $AWS_ROLE_ARN \
+                --role-session-name x-account \
+                --web-identity-token file://$AWS_WEB_IDENTITY_TOKEN_FILE \
+                --duration 1500 >/tmp/temp_creds.txt
 
-        export AWS_ACCESS_KEY_ID="$(cat /tmp/temp_creds.txt | jq -r ".Credentials.AccessKeyId")"
-        export AWS_SECRET_ACCESS_KEY="$(cat /tmp/temp_creds.txt | jq -r ".Credentials.SecretAccessKey")"
-        export AWS_SESSION_TOKEN="$(cat /tmp/temp_creds.txt | jq -r ".Credentials.SessionToken")"
+            export AWS_ACCESS_KEY_ID="$(cat /tmp/temp_creds.txt | jq -r ".Credentials.AccessKeyId")"
+            export AWS_SECRET_ACCESS_KEY="$(cat /tmp/temp_creds.txt | jq -r ".Credentials.SecretAccessKey")"
+            export AWS_SESSION_TOKEN="$(cat /tmp/temp_creds.txt | jq -r ".Credentials.SessionToken")"
 
-        rm /tmp/temp_creds.txt
+            rm /tmp/temp_creds.txt
 
+        fi
     fi
 
-
-    if ! aws secretsmanager get-secret-value --secret-id ${_SECRET_ID} | jq --raw-output '.SecretString' | jq -r .{_SECRET_KEY} ; then
+    if ! aws secretsmanager get-secret-value --secret-id ${_SECRET_ID} | jq --raw-output '.SecretString' | jq -r .{_SECRET_KEY}; then
         echo "Error while get secret from aws secrets manager!" >&2
         echo aws secretsmanager get-secret-value --secret-id "${_SECRET_ID}" | jq --raw-output '.SecretString' | jq -r ."{_SECRET_KEY}" >&2
         exit 1
